ai-governance-main/backend/Agents/requirements.txt

literalai
markdown-it-py
MarkupSafe
marshmallow
mcp
mdurl
monotonic
multidict
mypy_extensions
narwhals
nest-asyncio
numpy
openai
openai-agents
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-alephalpha
opentelemetry-instrumentation-anthropic
opentelemetry-instrumentation-bedrock
opentelemetry-instrumentation-chromadb
opentelemetry-instrumentation-cohere
opentelemetry-instrumentation-crewai
opentelemetry-instrumentation-google-generativeai
opentelemetry-instrumentation-groq
opentelemetry-instrumentation-haystack
opentelemetry-instrumentation-lancedb
opentelemetry-instrumentation-langchain
opentelemetry-instrumentation-llamaindex
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-marqo
opentelemetry-instrumentation-mcp
opentelemetry-instrumentation-milvus
opentelemetry-instrumentation-mistralai
opentelemetry-instrumentation-ollama
opentelemetry-instrumentation-openai
opentelemetry-instrumentation-openai-agents
opentelemetry-instrumentation-pinecone
opentelemetry-instrumentation-qdrant
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-replicate
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-sagemaker
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-threading
opentelemetry-instrumentation-together
opentelemetry-instrumentation-transformers
opentelemetry-instrumentation-urllib3
opentelemetry-instrumentation-vertexai
opentelemetry-instrumentation-watsonx
opentelemetry-instrumentation-weaviate
opentelemetry-instrumentation-writer
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-semantic-conventions-ai
opentelemetry-util-http
orjson
ormsgpack
packaging
pandas
pillow
portalocker
posthog
propcache
proto-plus
protobuf
pyarrow
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
PyJWT
pymongo
pypdf
python-dateutil
python-dotenv
python-engineio
python-multipart
python-socketio
pytz
PyYAML
qdrant-client
referencing
regex
requests
requests-toolbelt
rich
rpds-py
rsa
shapely
simple-websocket
six
smmap
sniffio
SQLAlchemy
sse-starlette
starlette
syncer
tabulate
tenacity
tiktoken
tokenizers
toml
tomli
tornado
tqdm
traceloop-sdk
types-requests
typing-inspect
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
vertexai
watchdog
watchfiles
wrapt
wsproto
xxhash
yarl
zipp
zstandard
